% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7036Tools.R
\name{.jst_find_file}
\alias{.jst_find_file}
\title{Internal: find a bare filename in Data/, data/, or working directory}
\usage{
.jst_find_file(filename)
}
\description{
Internal: find a bare filename in Data/, data/, or working directory
}
\keyword{internal}
