% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7036Tools.R
\name{.jst_get_search_dirs}
\alias{.jst_get_search_dirs}
\title{Internal: get the ordered list of directories to search for data files}
\usage{
.jst_get_search_dirs()
}
\description{
Internal: get the ordered list of directories to search for data files
}
\keyword{internal}
