% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7036Tools.R
\name{jdesc}
\alias{jdesc}
\title{Computes basic descriptive statistics (N, non-missing, min, max, mean, SD)
for one or more variables in a data frame. Prints a formatted table and
invisibly returns the underlying results as a data frame.}
\usage{
jdesc(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Unquoted variable names within \code{data}.}
}
\value{
Invisibly returns a data frame of SPSS-like descriptive statistics. Also prints a table.
}
\description{
Computes basic descriptive statistics (N, non-missing, min, max, mean, SD)
for one or more variables in a data frame. Prints a formatted table and
invisibly returns the underlying results as a data frame.
}
\examples{
jdesc(mtcars, mpg)
jdesc(mtcars, mpg, hp, wt)

}
